Glitch Fixed with the state machine
(Tom Hope Did help a tiiny weeny bit), Remeber set FOV in Degrees!!!

+++ gef-master/samples/create_mesh/NORMAL_TERRAIN_GEN_state.h
@@ -92,7 +92,7 @@ private:
 	//Update varibles
 	bool terrain_changed_;
 	bool terrain_shader_active_;
-	int fps_;
+	float fps_;
 
 
 	//RenderFunctions
